Rework the subprocess not to use fork
Fork approach is simple, but comes with several downsides.
The biggest one is that all requires are passed along to the
sub process, meaning things like 'optparse' can't be analyzed.
The new version spawns another Ruby process, that doesn't load
anything except for the inspected require.
Transport is switched to FIFO pipes.